EnvioRepository, ProvedorRepository

+++ PerfulandiaSPA/src/main/java/com/example/PerfulandiaSpa/repository/EnvioRepositoryJpa.java
@@ -4,7 +4,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.PerfulandiaSpa.model.Envios;
 
-public interface EnvioRepositoryJpa extends JpaRepository<Envios, Integer> {
-    // JpaRepository already provides methods like save(), findById(), and deleteById()
+
+
+
+
+
+public interface EnvioRepositoryJpa extends JpaRepository<Envios,Long>{
     public Envios findById(int id);
 }
